+++ CHANGELOG.md
@@ -17,6 +17,7 @@ Changes are grouped as follows
 ## [4.10.0] - 2022-10-11
 ### Added
 - Add `retrieve_latest` method to `cognite.client.sequences`
+- Add support for extending the expiration time of download links returned by `cognite.client.files.retrieve_download_urls()`
 
 ## [4.9.0] - 2022-10-10
 ### Added